:) 添加 bintray 配置
