+++ gradle.properties
@@ -15,3 +15,27 @@ org.gradle.jvmargs=-Xmx1536m
 # This option should only be used with decoupled projects. More details, visit
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 # org.gradle.parallel=true
+
+# 库的包名
+PROJ_GROUP=org.chx.kdroid
+# 库的ID
+PROJ_ARTIFACTID=kdroid
+# 库的版本
+PROJ_VERSION=1.0.0
+### 最后 gradle引用的形式就是 $PROJ_GROUP:$PROJ_ARTIFACTID:$PROJ_VERSION
+
+# 库名
+PROJ_NAME=KDroid
+# 库的项目主页
+PROJ_WEBSITEURL=https://github.com/7hens/KDroid
+# 问题跟踪地址
+PROJ_ISSUETRACKERURL=https://github.com/7hens/KDroid/issues
+# VCS 地址
+PROJ_VCSURL=https://github.com/7hens/KDroid.git
+# 库的简单描述
+PROJ_DESCRIPTION=Android library written in Kotlin
+
+# 开发者的信息, 可以随意
+DEVELOPER_ID=7hens
+DEVELOPER_NAME=Thens Wong
+DEVELOPER_EMAIL=thenstango@live.com
