--- library/src/main/AndroidManifest.xml
@@ -1,13 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-
-          package="org.chx.kandroid.jadapter"
->
-
-    <application android:allowBackup="true"
-                 android:label="@string/app_name"
-                 android:supportsRtl="true"
-    >
-
-    </application>
-
-</manifest>
