--- library/src/main/java/org/chx/kdroid/irefresh/DelegatedDataRefresher.kt
@@ -1,35 +1,0 @@
-package org.chx.kdroid.irefresh
-
-import org.chx.kdroid.candy.rx.EmptyObserver
-
-
-abstract class DelegatedDataRefresher<D>(protected var delegate: RefreshDelegate) : DataRefresher<D>() {
-    init {
-        delegate.whenLoad().subscribe(EmptyObserver.with {
-            when (it) {
-                RefreshDelegate.REFRESH -> refresh(true)
-                RefreshDelegate.LOAD_MORE -> loadMore(true)
-            }
-        })
-    }
-
-    fun endLoadingAnimation() {
-        delegate.isLoadingMore = false
-        delegate.isRefreshing = false
-    }
-
-    fun refresh(showsAnimation: Boolean) {
-        super.refresh()
-        delegate.isRefreshing = showsAnimation
-    }
-
-    fun loadMore(showsAnimation: Boolean) {
-        super.loadMore()
-        delegate.isLoadingMore = showsAnimation
-    }
-
-    override fun onDataSetChanged(convertedData: List<D>?) {
-        super.onDataSetChanged(convertedData)
-        endLoadingAnimation()
-    }
-}
