+++ sample/build.gradle
@@ -28,7 +28,7 @@ dependencies {
     })
     compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     
-    compile project(':library')
+    compile project(':kdroid')
 
     compile 'com.android.support:appcompat-v7:25.1.1'
     testCompile 'junit:junit:4.12'
